SER-72: Convert WAV to MP3 to play on site
Add an mp3 asset to EgraphAsset that is either initialized by EgraphActor or lazily initialized by audioMp3Url. This mp3 url now plays on the Egraph page.
With lazy initialization, we can be sure that all existing Egraphs are playable, but the first time a preexisting Egraph page is loaded, it will take longer as the wav is converted to mp3.
AudioConverter is a new object that executes the conversion, and it uses a new dependency called Xuggle. The new jar is 40MB, so permgen will need to be increased from the current 128MB.
The command to set a new PermGen is "bees app:deploy -a acme/test target/acme.war jvmPermSize=168", http://wiki.cloudbees.com/bin/view/RUN/JVM+PermGen+Space

Also, delete some unneeded Facebook registration example code.