SER-194: Admins should be able to self-serve print orders themselves now. There are new admin screens for PrintOrder that contain all the info needed to create prints as well as generate PNGs.
Evolution 24 is quite involved: CashTransaction is made the holder of Stripe data and billingPostalCode. Also, PrintOrder now holds pngUrl and amountPaidInCurrency (usually $45).
Nothing is deleted in the spirit of safe migrations. The only change in this evolution that is not backwards compatible is the refactor of CashTransaction's typeString to the new
enum pattern, but I will chance it since it is a low-risk change.

Changing EgraphPurchaseHandler so that Order.amountPaidInCurrency equals to product's price, PrintOrder.amountPaidInCurrency equal to 45 if appropriate,
and CashTransaction equal to their sum.

Death to PostBuyProductEndpoint! Post logic that is still used by PostBuyDemoProductEndpoint has been moved there, and
PostBuyProductEndpointTests has been deleted and its tests moved to EgraphPurchaseHandlerTests. EgraphPurchaseHandler is now its own file.